refactor: main page
- prefetch 적용
- MakeQuizDrawerDialog 컴포넌트 drawer 부분과 dialog 부분 분리
- MakeQuizDialog 내부 selector 부분 분리 (category/documents/quizCount)
- selector 분리 시 발생하는 ignoreId 부분 코드 중복을 최소화하기 위해 계산값을 useMemo로 저장
- types 디렉토리 하위 quiz.ts 생성 (MakeQuizDrawerDialog 컴포넌트에 있던 타입 정의)
- 추후 유지보수가 용이하도록 몇 가지 주석 추가
