+++ src/app/(routes)/(tabs)/main/page.tsx
@@ -1,10 +1,8 @@
-import icons from '@/constants/icons'
-import QuizBanner from './components/quiz-banner'
-import QuizMaker from './components/quiz-maker'
-import Achievements from './components/achievements'
-import { CommonLayout } from '@/shared/components/common-layout'
-import Image from 'next/image'
-import { RewordDialog } from './components/reward-alert'
+import { getQueryClient } from '@/shared/lib/tanstack-query/client'
+import { queries } from '@/shared/lib/tanstack-query/query-keys'
+import { currentMonth } from '@/shared/utils/date'
+import Main from '@/views/main'
+import { HydrationBoundary, dehydrate } from '@tanstack/react-query'
 
 interface Props {
   searchParams: {
@@ -12,33 +10,30 @@ interface Props {
   }
 }
 
-export default function Main({ searchParams }: Props) {
-  return (
-    <>
-      <CommonLayout
-        title={
-          <div className="flex gap-[8px]">
-            파워업 퀴즈
-            <div className="relative h-[21.67px] w-[25.05px] lg:h-[30px] lg:w-[34.69px]">
-              <Image src={icons.powerUpQuiz} fill alt="" />
-            </div>
-          </div>
-        }
-        mobileOptions={{
-          hasStars: true,
-          hasNotifications: true,
-        }}
-      >
-        <main className="flex flex-col px-[20px] pb-[40px] pt-[10px] lg:pt-[28px]">
-          <section className="mb-[56px] flex flex-col gap-[24px] lg:flex-row lg:gap-[10px]">
-            <QuizBanner />
-            <Achievements />
-          </section>
+const Page = async ({ searchParams: { reward } }: Props) => {
+  const queryClient = getQueryClient()
 
-          <QuizMaker />
-        </main>
-      </CommonLayout>
-      <RewordDialog reward={searchParams?.reward} />
-    </>
+  // 서버에서 미리 데이터 prefetching
+  await Promise.all([
+    queryClient.prefetchQuery(queries.category.list()),
+    queryClient.prefetchQuery(queries.quiz.today()),
+    queryClient.prefetchQuery(queries.quiz.monthAnswerRate({
+      categoryId: 0,
+      date: {
+        year: 2024,
+        month: currentMonth(),
+      },
+    })),
+  ]);
+
+  // 클라이언트로 데이터를 전달하기 위해 queryClient를 dehydrate 처리
+  const dehydratedState = dehydrate(queryClient)
+
+  return (
+    <HydrationBoundary state={dehydratedState}>
+      <Main reward={reward} />
+    </HydrationBoundary>
   )
 }
+
+export default Page
