+++ src/views/main/components/my-stars-drawer-dialog.tsx
